requirements.txt
langchain-google-genai==0.0.8
langchain-core==0.1.0
python-dotenv==1.0.0
